# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
        "/api/articles": {
            "get": {
                "summary": "Get a list of articles",
                "operationId": "d87f751e8f428fb968c2791aa22f017e",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Search keyword for articles",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "source",
                        "in": "query",
                        "description": "Filter by source",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Filter by date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Filter by category ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/articles/{id}": {
            "get": {
                "summary": "Get a specific article",
                "operationId": "dfa3f979aeffa953a3a227c721052741",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the article to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article retrieved successfully"
                    },
                    "404": {
                        "description": "Article not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user-personalized-feed/{userId}": {
            "get": {
                "summary": "Get personalized news feed for a user",
                "operationId": "d24068500cd6b11b4b546e7647b46b9f",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Personalized news feed retrieved successfully"
                    },
                    "404": {
                        "description": "User or preferences not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/categories": {
            "get": {
                "summary": "Get all categories",
                "operationId": "d0b3e412cb9fde72e014a0b12fc74f5a",
                "responses": {
                    "200": {
                        "description": "Categories retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "Category ID",
                                                "type": "integer"
                                            },
                                            "name": {
                                                "description": "Category Name",
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "summary": "Register a new user",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "User's name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User's email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User's password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "description": "Confirmation of the user's password. Must match the password field.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "summary": "Login a user",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User's email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User's password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Access token generated successfully"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "summary": "Logout a user",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User's email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Access token deleted successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/api/forgot-password": {
            "post": {
                "summary": "Send password reset link to the user",
                "operationId": "7af8817e88ba5874543ff38f188248f1",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User's email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Password reset link sent successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/api/reset-password": {
            "post": {
                "summary": "Reset user's password",
                "operationId": "00c0aad74830221c6da409a1355ec736",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Password reset token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User's email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "New password for the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "description": "Confirmation of the new password. Must match the password field.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Password reset successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/preferences": {
            "post": {
                "summary": "Set user preferences",
                "operationId": "91e95d87dc2e097db5d86e798b1632f4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "categories"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "ID of the user",
                                        "type": "integer"
                                    },
                                    "categories": {
                                        "description": "Array of category IDs",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "sources": {
                                        "description": "Array of source names",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "authors": {
                                        "description": "Array of author names",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User preferences set successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/preferences/{userId}": {
            "get": {
                "summary": "Get user preferences",
                "operationId": "087e2346f585c90f657755ab2b5d8520",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID of the user to get preferences for",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User preferences retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user_id": {
                                            "description": "User ID",
                                            "type": "integer"
                                        },
                                        "user_name": {
                                            "description": "User Name",
                                            "type": "string"
                                        },
                                        "categories": {
                                            "description": "List of category names",
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "sources": {
                                            "description": "List of sources",
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "authors": {
                                            "description": "List of authors",
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User or preferences not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}